* 'master' of https://github.com/supersede-project/feedback_gathering: (44 commits)
  Android Library: Cleanup and document code (#FEED-186)
  Android Library: Cleanup and document code (#FEED-186)
  Android Library: Cleanup and document code (#FEED-186)
  Android Library: Cleanup and document code (#FEED-186)
  Android Library: Adapt the client to the new API (#FEED-228) / Android Library: Add log messages for developers (#FEED-217)
  finished implmenetation of application scope authorization, changed urls of orchestrator
  added README to project
  class diagrams
  Android Library: Cleanup and document code (#FEED-186)
  Android Library: Add log messages for developers (#FEED-217)
  Android Library: Add log messages for developers (#FEED-217)
  Android Library: Add log messages for developers (#FEED-217)
  Android Library: Add log messages for developers (#FEED-217)
  Android Library: As a Developer I can trigger a specific pull configuration (#FEED-220)
  started implementation of application level user authorization
  Android Library: As a Developer I can set the maximum number of text annotations (#FEED-226)
  implemented and tested FeedbackStatusController and StatusOptionController
  master settings
  some adjustments
  the selected object moves into foreground
  ...

+++ repository/src/ch/uzh/ifi/feedback/repository/model/Feedback.java
@@ -4,12 +4,12 @@ import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
 
-import ch.uzh.ifi.feedback.library.rest.Service.ItemBase;
 import ch.uzh.ifi.feedback.library.rest.annotations.DbAttribute;
 import ch.uzh.ifi.feedback.library.rest.annotations.DbIgnore;
+import ch.uzh.ifi.feedback.library.rest.annotations.Id;
+import ch.uzh.ifi.feedback.library.rest.annotations.NotNull;
 import ch.uzh.ifi.feedback.library.rest.annotations.Serialize;
-import ch.uzh.ifi.feedback.library.rest.validation.Id;
-import ch.uzh.ifi.feedback.library.rest.validation.NotNull;
+import ch.uzh.ifi.feedback.library.rest.service.ItemBase;
 import ch.uzh.ifi.feedback.repository.serialization.FeedbackSerializationService;
 
 @Serialize(FeedbackSerializationService.class)
@@ -22,7 +22,7 @@ public class Feedback extends ItemBase<Feedback> {
 
 	@NotNull
 	@DbAttribute("application_id")
-	private long applicationId;
+	private Integer applicationId;
 
 	@DbAttribute("user_identification")
 	@NotNull
@@ -101,7 +101,7 @@ public class Feedback extends ItemBase<Feedback> {
 		this.id = id;
 	}
 
-	public long getApplicationId() {
+	public Integer getApplicationId() {
 		return applicationId;
 	}
